requirements.txt
selenium
pyautogui
beautifulsoup4
flask
gunicorn
psycopg2-binary

